Bug: 32 bit int shifted by 32 bits in org.eclipse.pde.internal.swt.tools.IconExe.read8(RandomAccessFile)
The code performs shift of a 32 bit int by a constant amount outside the
range -31..31. The effect of this is to use the lower 5 bits of the
integer value to decide how much to shift by (e.g., shifting by 40 bits
is the same as shifting by 8 bits, and shifting by 32 bits is the same
as shifting by zero bits). This probably isn't what was expected, and it
is at least confusing.

Rank: Scariest (1), confidence: High
Pattern: ICAST_BAD_SHIFT_AMOUNT
Type: BSHIFT, Category: CORRECTNESS (Correctness)

+++ bundles/org.eclipse.equinox.p2.publisher.eclipse/src/org/eclipse/pde/internal/swt/tools/IconExe.java
@@ -748,14 +748,14 @@ public class IconExe {
 	}
 
 	static long read8(RandomAccessFile raf) throws IOException {
-		int b0 = raf.readByte() & 0xFF;
-		int b1 = raf.readByte() & 0xFF;
-		int b2 = raf.readByte() & 0xFF;
-		int b3 = raf.readByte() & 0xFF;
-		int b4 = raf.readByte() & 0xFF;
-		int b5 = raf.readByte() & 0xFF;
-		int b6 = raf.readByte() & 0xFF;
-		int b7 = raf.readByte() & 0xFF;
+		long b0 = raf.readByte() & 0xFF;
+		long b1 = raf.readByte() & 0xFF;
+		long b2 = raf.readByte() & 0xFF;
+		long b3 = raf.readByte() & 0xFF;
+		long b4 = raf.readByte() & 0xFF;
+		long b5 = raf.readByte() & 0xFF;
+		long b6 = raf.readByte() & 0xFF;
+		long b7 = raf.readByte() & 0xFF;
 		return b7 << 56 | b6 << 48 | b5 << 40 | b4 << 32 | b3 << 24 | b2 << 16 | b1 << 8 | b0;
 	}
